binja_frontend: Handle dupe but different symbols
Binary Ninja apparently allows multiple symbols to exist for the same
address, and the ordering in the list we get is nondeterministic.

To prevent rename spam, we individually verify each symbol is the
currently used one via `get_symbol_at`. This is probably slow, but I
don't know what else to do.